Fix `slide_basic` example
